--- a/api/base/filters.py
+++ b/api/base/filters.py
@@ -14,11 +14,11 @@ from api.base.serializers import RelationshipField, ShowIfVersion, TargetField
 from dateutil import parser as date_parser
 from django.core.exceptions import ValidationError
 from django.db.models import QuerySet as DjangoQuerySet
-from django.db.models import Q
+from django.db.models import Q, Exists, OuterRef
 from modularodm.query import queryset as modularodm_queryset
 from rest_framework import serializers as ser
 from rest_framework.filters import OrderingFilter
-from osf.models import Subject, PreprintProvider
+from osf.models import Subject, PreprintProvider, Node
 from osf.models.base import GuidMixin
 from reviews.workflow import States
 
@@ -498,7 +498,7 @@ class PreprintFilterMixin(ListFilterMixin):
                 operation['op'] = 'iexact'
 
     def preprints_queryset(self, base_queryset, auth_user, allow_contribs=True):
-        default_query = Q(node__isnull=False, node__is_deleted=False)
+        sub_qs = Node.objects.filter(preprints=OuterRef('pk'), is_deleted=False)
         no_user_query = Q(is_published=True, node__is_public=True)
 
         if auth_user:
@@ -506,10 +506,10 @@ class PreprintFilterMixin(ListFilterMixin):
             reviews_user_query = Q(node__is_public=True, provider__in=get_objects_for_user(auth_user, 'view_submissions', PreprintProvider))
             if allow_contribs:
                 contrib_user_query = ~Q(reviews_state=States.INITIAL.value) & Q(node__contributor__user_id=auth_user.id, node__contributor__read=True)
-                query = default_query & (no_user_query | contrib_user_query | admin_user_query | reviews_user_query)
+                query = (no_user_query | contrib_user_query | admin_user_query | reviews_user_query)
             else:
-                query = default_query & (no_user_query | admin_user_query | reviews_user_query)
+                query = (no_user_query | admin_user_query | reviews_user_query)
         else:
-            query = default_query & no_user_query
+            query = no_user_query
 
-        return base_queryset.filter(query)
+        return base_queryset.annotate(default=Exists(sub_qs)).filter(Q(default=True) & query)