--- a/api/base/utils.py
+++ b/api/base/utils.py
@@ -13,7 +13,7 @@ from api.base.exceptions import Gone, UserGone
 from framework.auth import Auth
 from framework.auth.cas import CasResponse
 from framework.auth.oauth_scopes import ComposedScopes, normalize_scopes
-from osf.models import OSFUser, Contributor
+from osf.models import OSFUser, Contributor, Node, Registration
 from osf.models.base import GuidMixin
 from osf.modm_compat import to_django_query
 from osf.utils.requests import check_select_for_update
@@ -153,11 +153,15 @@ def waterbutler_url_for(request_type, provider, path, node_id, token, obj_args=N
     url.args.update(query)
     return url.url
 
+def check_model_cls(model_cls):
+    assert model_cls is Node or model_cls is Registration
 
 def default_node_list_queryset(model_cls):
+    check_model_cls(model_cls)
     return model_cls.objects.filter(is_deleted=False)
 
 def default_node_permission_queryset(user, model_cls):
+    check_model_cls(model_cls)
     if user.is_anonymous:
         return model_cls.objects.filter(is_public=True)
     sub_qs = Contributor.objects.filter(node=OuterRef('pk'), user__id=user.id, read=True)
@@ -168,7 +172,6 @@ def default_node_list_permission_queryset(user, model_cls):
     # If get_roots() is called on default_node_list_qs & default_node_permission_qs,
     # Django's alaising will break and the resulting QS will be empty and you will be sad.
     return default_node_permission_queryset(user, model_cls) & default_node_list_queryset(model_cls)
-
 
 def extend_querystring_params(url, params):
     scheme, netloc, path, query, _ = urlparse.urlsplit(url)