--- a/api/preprint_providers/views.py
+++ b/api/preprint_providers/views.py
@@ -206,8 +206,9 @@ class PreprintProviderPreprintList(JSONAPIBaseView, generics.ListAPIView, Prepri
 
         if auth_user:
             contrib_user_query = Q(is_published=True, node__contributor__user_id=auth_user.id, node__contributor__read=True)
-            sub_qs = Contributor.objects.filter(node=OuterRef('pk'), user_id=auth_user.id, admin=True)
-            return default_qs.annotate(admin_user=Exists(sub_qs)).filter(Q(admin_user=True) | contrib_user_query | no_user_query)
+            admin_subquery = Contributor.objects.filter(node=OuterRef('pk'), user_id=auth_user.id, admin=True)
+            node_subquery = AbstractNode.objects.annotate(contrib=Exists(admin_subquery)).filter(preprints=OuterRef('pk'), contrib=True)
+            return default_qs.annotate(node_present=Exists(node_subquery)).filter(Q(node_present=True) | contrib_user_query | no_user_query)
         return default_qs.filter(no_user_query)
 
     # overrides ListAPIView