## Purpose

DISTINCT ON is slow and not necessary (can be worked around by using subqueries). Change .distinct() calls to use subqueries instead.

## Changes
- Revert Casey's commit 72e5f1fd59b5f0feea3f43ed97f3ee70cdec2cbf
- Remove distinct usages on NodeList, RegistrationList, UserNodes (distinct usage orig. added here: https://github.com/CenterForOpenScience/osf.io/pull/7682) , and User Registrations following Casey's work

### Remove distinct usage in favor of subqueries on:
#### API v2 Node Files ListView
- Distinct usage originally added after this comment https://github.com/CenterForOpenScience/osf.io/pull/6770#pullrequestreview-18247993).
- Can test that duplicates are still filtered by filtering tags on node files. Add tags with different casing but same letters.  So give file 'pq', and 'PQ' tags, and then filter`/v2/nodes/3mwvb/files/osfstorage/?filter[tags]=pq`
- This subquery used here is very redundant, but removes the need for `distinct on`.  Timing seems to be similar. 

#### API v2 Node Preprints ListView
- Distinct usage originally added here https://github.com/CenterForOpenScience/osf.io/pull/7162

#### API v2 Preprint Provider Preprint ListView

#### API v2 Preprints ListView
#### API v2 User Preprints ListView
#### ~populate new and noteworthy nodes script~
I think this is a good usage of `distinct`. Here we are not trying to filter out duplicate logs.  The script is trying to find the greatest number of unique log actions.  So it is filtering out logs with same action, not duplicate logs.
#### ~send_preprint_unreg_contributor_emails script~ 
- Script should only run once, so not necessary to change
#### get_metaschemas method in `drafts.py`
- Distinct does not seem to be needed here. Orginally added https://github.com/CenterForOpenScience/osf.io/pull/7703
- Original `distinct` was being called on a values_list of names.  There are metaschemas with duplicate names, so it made sense why it was being used there.
```
list(MetaSchema.objects.all().values_list('name', flat=True).distinct())
```
- However, MetaSchema.objects.filter(active=True).filter(schema_version=LATEST_SCHEMA_VERSION) is just filtering on MetaSchema fields, no related fields, so this shouldn't be producing duplicates in the first place.

## Benchmarking
Before 
### Averages, stdev given in Elapsed Time (ms) to complete request
  | # queries | STDEV, 10 trials | AVG, 10 trials
-- | -- | -- | --
Node List | 78 | 122.00 | 1,591.70
Registration List | 33 | 131.07 | 682.03
UserNodes | 58 | 121.20 | 1,279.09
UserRegistrations | 27 | 90.64 | 613.77
NodeFiles ListView | 17 | 75.79 | 389.51
Node Preprints List View (ackvb) | 18 | 100.97 | 390.32
Preprint Provider Preprint List View | 63 | 112.91 | 1,140.92
Preprints List View | 72 | 120.37 | 1,274.17
User Preprints List View | 72 | 213.69 | 1,385.01



After
### Averages, stdev given in Elapsed Time (ms) to complete request

  | # queries | STDEV, 10 trials | AVG, 10 trials
-- | -- | -- | --
NodeList | 78 | 197.51 | 1,656.42
Registration List | 33 | 84.39 | 677.85
UserNodes | 58 | 177.25 | 1,276.89
UserRegistrations | 27 | 123.01 | 659.10
NodeFiles ListView | 17 | 104.83 | 427.34
Node Preprints List View (ackvb) | 18 | 80.66 | 482.50
Preprint Provider Preprint List View | 66 | 84.34 | 1,332.44
Preprints List View | 75 | 161.97 | 1,508.77
User Preprints List View | 75 | 198.08 | 1,464.06



## My Notes
Example using subquery to get queryset of files with certain tag.

``` python
from osf.models import Tag
from django.db.models import Exists, OuterRef
sub_qs =  Tag.objects.filter(basefilenode_tagged=OuterRef('pk'), name__iexact='<tag_name>')
return queryset.annotate(tag=Exists(sub_qs)).filter(Q(tag=True))
```
## Side effects

<!--Any possible side effects? -->


## Ticket

https://openscience.atlassian.net/browse/OSF-8639