@pattisdr 2017-10-26
@caseyrollins thanks for the review! I added the benchmarking in the description; I didn't notice much of a difference testing locally.