@pattisdr 2017-11-13
Thanks @sloria, will address later this afternoon